pyproject.toml
[tool.poetry]
name = "mlopdracht"
version = "0.1.0"
description = ""
authors = ["R.Grouls <rgrouls@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
torch = { url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp39-cp39-linux_x86_64.whl"}
loguru = "^0.7.0"
tqdm = "^4.65.0"
numpy = "^1.24.3"
requests = "^2.30.0"
pydantic = "^1.10.7"
jupyter = "^1.0.0"
types-requests = "^2.30.0.0"
quarto = "^0.1.0"
gin-config = "^0.5.0"
mlflow = "^2.3.2"
ray = "^2.4.0"
tensorboard = "^2.13.0"
hpbandster = "^0.7.4"
configspace = "^0.6.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pep8-naming = "^0.13.3"
flake8-annotations = "^3.0.1"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
